requirements.txt
psycopg2-binary
pandas
streamlit
sqlalchemy
altair